strange bug
phenomenon:

program seems to be trapped in dead loop when PU >= 32, but it run
normally when I add lots of print in it

analysis:

This bug has something to do with program's running speed. Because it
disappeared when print slowed it down.

Later on, I test it again under print removal. Mark current version as a
temporary successful one

Signed-off-by: virgilshi <2338052892@qq.com>